frame/inspect.rs: extract remaining inspect cluster (#988 Phase 2b)
Move 15 more parse / inspect fns from frame/mod.rs into frame/
inspect.rs:

  authoritative_forward_ports, live_frame_ports, live_frame_ports_bytes,
  live_frame_ports_from_meta_bytes, forward_tuple_mismatch_reason,
  parse_session_flow, parse_session_flow_from_bytes,
  parse_session_flow_from_frame, parse_session_flow_from_meta,
  parse_ipv4_session_flow_from_frame, decode_frame_summary,
  parse_zone_encoded_fabric_ingress,
  parse_zone_encoded_fabric_ingress_from_frame,
  parse_packet_destination_from_frame, try_parse_metadata.

frame/mod.rs production LOC (excluding mod tests):
  2,487 → 1,993  (UNDER the 2,000 threshold)
inspect.rs: 349 → 861 LOC.

Visibility: fns called from outside the frame/ tree
(authoritative_forward_ports, decode_frame_summary,
forward_tuple_mismatch_reason, frame_has_tcp_rst,
parse_session_flow, try_parse_metadata) get re-exported at
pub(in crate::afxdp); the rest stay pub(super).

Inspect.rs uses `use super::*;` to bring in the broader type
surface (MmapArea, XdpDesc, IpAddr, SessionKey, ForwardingState,
ForwardPacketMeta, etc.) that the bigger inspect fns need —
explicit imports here would be a pages-long list with no
disambiguation benefit.

Per docs/refactor/988-frame-decomp.md (origin/research/988-frame-
design). Phase 0/1 (checksum) in #1019/#1021, Phase 2a (10 leaf
fns) in #1026.

build clean (cargo build --release), 865 Rust tests pass.
