+++ userspace-dp/src/afxdp/frame/inspect.rs
@@ -1,12 +1,12 @@
-//! Header inspection / parsing helpers — pure read-only fns over
+//! Header inspection / parsing helpers — read-only fns over
 //! Ethernet/IPv4/IPv6/TCP/UDP/ICMP byte buffers. No mutation.
 //!
-//! Phase 2 split out of `frame.rs` per #988. Larger inspect fns
-//! (parse_session_flow_*, authoritative_forward_ports,
-//! try_parse_metadata) remain in mod.rs for now and will land in a
-//! follow-on phase to keep this diff reviewable.
+//! Phase 2 split out of `frame.rs` per #988. The inspect cluster
+//! covers raw header parsing (frame_l3_offset, parse_session_flow_*,
+//! decode_frame_summary, etc.) plus session-key / fabric-tag readers
+//! that operate on a frame slice without mutating it.
 
-use super::{PROTO_ICMP, PROTO_ICMPV6, PROTO_TCP, PROTO_UDP, SessionFlow, UserspaceDpMeta};
+use super::*;
 
 /// Check if a frame contains a TCP RST flag.
 pub(in crate::afxdp) fn frame_has_tcp_rst(frame: &[u8]) -> bool {
@@ -350,4 +350,512 @@ pub(in crate::afxdp) fn parse_flow_ports(
         }
         _ => None,
     }
+}
+
+#[cfg_attr(not(test), allow(dead_code))]
+pub(in crate::afxdp) fn authoritative_forward_ports(
+    frame: &[u8],
+    meta: UserspaceDpMeta,
+    flow: Option<&SessionFlow>,
+) -> Option<(u16, u16)> {
+    if !matches!(meta.protocol, PROTO_TCP | PROTO_UDP) {
+        return None;
+    }
+    if let Some(flow_ports) = flow.and_then(|flow| {
+        if flow.forward_key.src_port != 0 && flow.forward_key.dst_port != 0 {
+            Some((flow.forward_key.src_port, flow.forward_key.dst_port))
+        } else {
+            None
+        }
+    }) {
+        return Some(flow_ports);
+    }
+    let meta_ports = if meta.flow_src_port != 0 && meta.flow_dst_port != 0 {
+        Some((meta.flow_src_port, meta.flow_dst_port))
+    } else {
+        None
+    };
+    let frame_ports = live_frame_ports_from_meta_bytes(frame, meta);
+    frame_ports.or(meta_ports)
+}
+
+#[allow(dead_code)]
+pub(in crate::afxdp) fn live_frame_ports(
+    area: &MmapArea,
+    desc: XdpDesc,
+    meta: UserspaceDpMeta,
+) -> Option<(u16, u16)> {
+    if !matches!(meta.protocol, PROTO_TCP | PROTO_UDP) {
+        return None;
+    }
+    let frame = area.slice(desc.addr as usize, desc.len as usize)?;
+    live_frame_ports_from_meta_bytes(frame, meta)
+}
+
+#[inline(always)]
+pub(in crate::afxdp) fn live_frame_ports_from_meta_bytes(
+    frame: &[u8],
+    meta: impl Into<ForwardPacketMeta>,
+) -> Option<(u16, u16)> {
+    let meta = meta.into();
+    if !matches!(meta.protocol, PROTO_TCP | PROTO_UDP) {
+        return None;
+    }
+    let l4 = meta.l4_offset as usize;
+    if l4 != 0
+        && let Some(ports) = parse_flow_ports(frame, l4, meta.protocol)
+    {
+        return Some(ports);
+    }
+    live_frame_ports_bytes(frame, meta.addr_family, meta.protocol)
+}
+
+pub(in crate::afxdp) fn live_frame_ports_bytes(
+    frame: &[u8],
+    addr_family: u8,
+    protocol: u8,
+) -> Option<(u16, u16)> {
+    if !matches!(protocol, PROTO_TCP | PROTO_UDP) {
+        return None;
+    }
+    let l4 = frame_l4_offset(frame, addr_family)?;
+    parse_flow_ports(frame, l4, protocol)
+}
+
+pub(in crate::afxdp) fn forward_tuple_mismatch_reason(
+    source_ports: Option<(u16, u16)>,
+    expected_ports: Option<(u16, u16)>,
+    built_ports: Option<(u16, u16)>,
+) -> Option<String> {
+    let expected = expected_ports.or(source_ports)?;
+    let built = built_ports?;
+    if built == expected {
+        return None;
+    }
+    let source = source_ports.unwrap_or((0, 0));
+    Some(format!(
+        "forward_tuple_mismatch:src={}:{} expected={}:{} built={}:{}",
+        source.0, source.1, expected.0, expected.1, built.0, built.1
+    ))
+}
+
+pub(in crate::afxdp) fn parse_session_flow_from_bytes(
+    frame: &[u8],
+    meta: UserspaceDpMeta,
+) -> Option<SessionFlow> {
+    let meta_flow = parse_session_flow_from_meta(meta);
+    // Fast path: for TCP/UDP with complete metadata tuple, use meta directly
+    // without parsing the frame. This avoids extra L3/L4 parsing for the
+    // common established-flow case.
+    if matches!(meta.protocol, PROTO_TCP | PROTO_UDP)
+        && let Some(meta_flow) = meta_flow.as_ref()
+        && metadata_tuple_complete(meta, meta_flow)
+    {
+        return Some(meta_flow.clone());
+    }
+
+    let frame_flow = if matches!(meta.addr_family as i32, libc::AF_INET) {
+        parse_ipv4_session_flow_from_frame(frame, meta)
+    } else {
+        parse_session_flow_from_frame(frame, meta)
+    };
+
+    if let Some(meta_flow) = meta_flow
+        && metadata_tuple_complete(meta, &meta_flow)
+    {
+        if let Some(ref frame_flow) = frame_flow {
+            if frame_flow.src_ip == meta_flow.src_ip && frame_flow.dst_ip == meta_flow.dst_ip {
+                return Some(meta_flow);
+            }
+            return Some(frame_flow.clone());
+        }
+        return Some(meta_flow);
+    }
+
+    if let Some(flow) = frame_flow {
+        return Some(flow);
+    }
+
+    let l3 = meta.l3_offset as usize;
+    let l4 = meta.l4_offset as usize;
+    match meta.addr_family as i32 {
+        libc::AF_INET => {
+            if frame.len() < l3 + 20 || frame.len() < l4 {
+                return None;
+            }
+            let src_ip = IpAddr::V4(Ipv4Addr::new(
+                frame[l3 + 12],
+                frame[l3 + 13],
+                frame[l3 + 14],
+                frame[l3 + 15],
+            ));
+            let dst_ip = IpAddr::V4(Ipv4Addr::new(
+                frame[l3 + 16],
+                frame[l3 + 17],
+                frame[l3 + 18],
+                frame[l3 + 19],
+            ));
+            let (src_port, dst_port) = parse_flow_ports(frame, l4, meta.protocol)?;
+            Some(SessionFlow {
+                src_ip,
+                dst_ip,
+                forward_key: SessionKey {
+                    addr_family: meta.addr_family,
+                    protocol: meta.protocol,
+                    src_ip,
+                    dst_ip,
+                    src_port,
+                    dst_port,
+                },
+            })
+        }
+        libc::AF_INET6 => {
+            if frame.len() < l3 + 40 || frame.len() < l4 {
+                return None;
+            }
+            let src_ip = IpAddr::V6(Ipv6Addr::from(
+                <[u8; 16]>::try_from(&frame[l3 + 8..l3 + 24]).ok()?,
+            ));
+            let dst_ip = IpAddr::V6(Ipv6Addr::from(
+                <[u8; 16]>::try_from(&frame[l3 + 24..l3 + 40]).ok()?,
+            ));
+            let (src_port, dst_port) = parse_flow_ports(frame, l4, meta.protocol)?;
+            Some(SessionFlow {
+                src_ip,
+                dst_ip,
+                forward_key: SessionKey {
+                    addr_family: meta.addr_family,
+                    protocol: meta.protocol,
+                    src_ip,
+                    dst_ip,
+                    src_port,
+                    dst_port,
+                },
+            })
+        }
+        _ => None,
+    }
+}
+
+#[cfg_attr(not(test), allow(dead_code))]
+pub(in crate::afxdp) fn parse_session_flow(
+    area: &MmapArea,
+    desc: XdpDesc,
+    meta: UserspaceDpMeta,
+) -> Option<SessionFlow> {
+    let frame = area.slice(desc.addr as usize, desc.len as usize)?;
+    parse_session_flow_from_bytes(frame, meta)
+}
+
+/// Decode an Ethernet frame into a human-readable summary showing IP src/dst,
+/// TCP/UDP ports, TCP flags, and checksums. For debugging packet forwarding.
+pub(in crate::afxdp) fn decode_frame_summary(frame: &[u8]) -> String {
+    let l3 = match frame_l3_offset(frame) {
+        Some(off) => off,
+        None => return String::new(),
+    };
+    let ip = &frame[l3..];
+    if ip.len() < 20 {
+        return String::new();
+    }
+    let version = ip[0] >> 4;
+    if version == 4 {
+        let ihl = ((ip[0] & 0x0f) as usize) * 4;
+        let total_len = u16::from_be_bytes([ip[2], ip[3]]);
+        let protocol = ip[9];
+        let ip_csum = u16::from_be_bytes([ip[10], ip[11]]);
+        let src = format!("{}.{}.{}.{}", ip[12], ip[13], ip[14], ip[15]);
+        let dst = format!("{}.{}.{}.{}", ip[16], ip[17], ip[18], ip[19]);
+        let ttl = ip[8];
+        if matches!(protocol, PROTO_TCP | PROTO_UDP) && ip.len() >= ihl + 8 {
+            let l4 = &ip[ihl..];
+            let sport = u16::from_be_bytes([l4[0], l4[1]]);
+            let dport = u16::from_be_bytes([l4[2], l4[3]]);
+            if protocol == PROTO_TCP && ip.len() >= ihl + 20 {
+                let seq = u32::from_be_bytes([l4[4], l4[5], l4[6], l4[7]]);
+                let ack = u32::from_be_bytes([l4[8], l4[9], l4[10], l4[11]]);
+                let flags = l4[13];
+                let tcp_csum = u16::from_be_bytes([l4[16], l4[17]]);
+                let flag_str = tcp_flags_str(flags);
+                format!(
+                    "IPv4 {}:{} -> {}:{} TCP [{flag_str}] seq={seq} ack={ack} ttl={ttl} ip_csum={ip_csum:#06x} tcp_csum={tcp_csum:#06x} ip_len={total_len}",
+                    src, sport, dst, dport,
+                )
+            } else if protocol == PROTO_UDP {
+                let udp_csum = u16::from_be_bytes([l4[6], l4[7]]);
+                format!(
+                    "IPv4 {}:{} -> {}:{} UDP ttl={ttl} ip_csum={ip_csum:#06x} udp_csum={udp_csum:#06x} ip_len={total_len}",
+                    src, sport, dst, dport,
+                )
+            } else {
+                format!(
+                    "IPv4 {} -> {} proto={protocol} ttl={ttl} ip_len={total_len}",
+                    src, dst
+                )
+            }
+        } else {
+            format!(
+                "IPv4 {} -> {} proto={protocol} ttl={ttl} ip_len={total_len}",
+                src, dst
+            )
+        }
+    } else if version == 6 && ip.len() >= 40 {
+        let payload_len = u16::from_be_bytes([ip[4], ip[5]]);
+        let next_header = ip[6];
+        let hop_limit = ip[7];
+        let src = std::net::Ipv6Addr::from(<[u8; 16]>::try_from(&ip[8..24]).unwrap_or([0; 16]));
+        let dst = std::net::Ipv6Addr::from(<[u8; 16]>::try_from(&ip[24..40]).unwrap_or([0; 16]));
+        if matches!(next_header, PROTO_TCP | PROTO_UDP) && ip.len() >= 48 {
+            let l4 = &ip[40..];
+            let sport = u16::from_be_bytes([l4[0], l4[1]]);
+            let dport = u16::from_be_bytes([l4[2], l4[3]]);
+            if next_header == PROTO_TCP && ip.len() >= 60 {
+                let flags = l4[13];
+                let flag_str = tcp_flags_str(flags);
+                format!(
+                    "IPv6 [{src}]:{sport} -> [{dst}]:{dport} TCP [{flag_str}] hop={hop_limit} pl={payload_len}"
+                )
+            } else {
+                format!(
+                    "IPv6 [{src}]:{sport} -> [{dst}]:{dport} proto={next_header} hop={hop_limit} pl={payload_len}"
+                )
+            }
+        } else {
+            format!("IPv6 [{src}] -> [{dst}] proto={next_header} hop={hop_limit} pl={payload_len}")
+        }
+    } else {
+        String::new()
+    }
+}
+
+pub(in crate::afxdp) fn parse_session_flow_from_frame(
+    frame: &[u8],
+    meta: UserspaceDpMeta,
+) -> Option<SessionFlow> {
+    match meta.addr_family as i32 {
+        libc::AF_INET => parse_ipv4_session_flow_from_frame(frame, meta),
+        libc::AF_INET6 => {
+            let l3 = match meta.l3_offset {
+                14 | 18
+                    if frame
+                        .get(meta.l3_offset as usize)
+                        .is_some_and(|byte| (byte >> 4) == 6) =>
+                {
+                    meta.l3_offset as usize
+                }
+                _ => frame_l3_offset(frame)?,
+            };
+            let meta_rel = meta.l4_offset.wrapping_sub(meta.l3_offset) as usize;
+            let l4 = if meta_rel >= 40 && meta.l4_offset > meta.l3_offset {
+                l3.checked_add(meta_rel)?
+            } else {
+                frame_l4_offset(frame, meta.addr_family)?
+            };
+            if frame.len() < l3 + 40 || frame.len() < l4 {
+                return None;
+            }
+            let src_ip = IpAddr::V6(Ipv6Addr::from(
+                <[u8; 16]>::try_from(&frame[l3 + 8..l3 + 24]).ok()?,
+            ));
+            let dst_ip = IpAddr::V6(Ipv6Addr::from(
+                <[u8; 16]>::try_from(&frame[l3 + 24..l3 + 40]).ok()?,
+            ));
+            let (src_port, dst_port) = parse_flow_ports(frame, l4, meta.protocol)?;
+            Some(SessionFlow {
+                src_ip,
+                dst_ip,
+                forward_key: SessionKey {
+                    addr_family: meta.addr_family,
+                    protocol: meta.protocol,
+                    src_ip,
+                    dst_ip,
+                    src_port,
+                    dst_port,
+                },
+            })
+        }
+        _ => None,
+    }
+}
+
+pub(in crate::afxdp) fn parse_session_flow_from_meta(meta: UserspaceDpMeta) -> Option<SessionFlow> {
+    let (src_ip, dst_ip) = match meta.addr_family as i32 {
+        libc::AF_INET => {
+            let src = meta.flow_src_addr.get(..4)?;
+            let dst = meta.flow_dst_addr.get(..4)?;
+            (
+                IpAddr::V4(Ipv4Addr::new(src[0], src[1], src[2], src[3])),
+                IpAddr::V4(Ipv4Addr::new(dst[0], dst[1], dst[2], dst[3])),
+            )
+        }
+        libc::AF_INET6 => (
+            IpAddr::V6(Ipv6Addr::from(meta.flow_src_addr)),
+            IpAddr::V6(Ipv6Addr::from(meta.flow_dst_addr)),
+        ),
+        _ => return None,
+    };
+    if src_ip.is_unspecified() || dst_ip.is_unspecified() {
+        return None;
+    }
+    Some(SessionFlow {
+        src_ip,
+        dst_ip,
+        forward_key: SessionKey {
+            addr_family: meta.addr_family,
+            protocol: meta.protocol,
+            src_ip,
+            dst_ip,
+            src_port: meta.flow_src_port,
+            dst_port: meta.flow_dst_port,
+        },
+    })
+}
+
+pub(in crate::afxdp) fn parse_ipv4_session_flow_from_frame(
+    frame: &[u8],
+    meta: UserspaceDpMeta,
+) -> Option<SessionFlow> {
+    let l3 = match meta.l3_offset {
+        14 | 18
+            if frame
+                .get(meta.l3_offset as usize)
+                .is_some_and(|byte| (byte >> 4) == 4) =>
+        {
+            meta.l3_offset as usize
+        }
+        _ => frame_l3_offset(frame)?,
+    };
+    if frame.len() < l3 + 20 {
+        return None;
+    }
+    let ihl = usize::from(frame[l3] & 0x0f) * 4;
+    if ihl < 20 || frame.len() < l3 + ihl {
+        return None;
+    }
+    let protocol = frame[l3 + 9];
+    let parsed_l4 = l3 + ihl;
+    let l4 = if meta.l4_offset > meta.l3_offset && meta.l4_offset as usize == parsed_l4 {
+        meta.l4_offset as usize
+    } else {
+        parsed_l4
+    };
+    if frame.len() < l4 {
+        return None;
+    }
+    let src_ip = IpAddr::V4(Ipv4Addr::new(
+        frame[l3 + 12],
+        frame[l3 + 13],
+        frame[l3 + 14],
+        frame[l3 + 15],
+    ));
+    let dst_ip = IpAddr::V4(Ipv4Addr::new(
+        frame[l3 + 16],
+        frame[l3 + 17],
+        frame[l3 + 18],
+        frame[l3 + 19],
+    ));
+    let (src_port, dst_port) = parse_flow_ports(frame, l4, protocol)?;
+    Some(SessionFlow {
+        src_ip,
+        dst_ip,
+        forward_key: SessionKey {
+            addr_family: meta.addr_family,
+            protocol,
+            src_ip,
+            dst_ip,
+            src_port,
+            dst_port,
+        },
+    })
+}
+
+#[cfg_attr(not(test), allow(dead_code))]
+pub(in crate::afxdp) fn parse_zone_encoded_fabric_ingress(
+    area: &MmapArea,
+    desc: XdpDesc,
+    meta: UserspaceDpMeta,
+    forwarding: &ForwardingState,
+) -> Option<u16> {
+    let frame = area.slice(desc.addr as usize, desc.len as usize)?;
+    parse_zone_encoded_fabric_ingress_from_frame(frame, meta, forwarding)
+}
+
+/// #919/#922: returns the encoded zone ID (u8 → u16) directly, no
+/// `zone_id_to_name` lookup or `String` clone. Callers that need a
+/// name resolve via `forwarding.zone_id_to_name` on the slow path.
+pub(in crate::afxdp) fn parse_zone_encoded_fabric_ingress_from_frame(
+    frame: &[u8],
+    meta: UserspaceDpMeta,
+    forwarding: &ForwardingState,
+) -> Option<u16> {
+    if !ingress_is_fabric(forwarding, meta.ingress_ifindex as i32) {
+        return None;
+    }
+    if frame.len() < 12 {
+        return None;
+    }
+    if frame[6] != 0x02
+        || frame[7] != 0xbf
+        || frame[8] != 0x72
+        || frame[9] != FABRIC_ZONE_MAC_MAGIC
+        || frame[10] != 0x00
+    {
+        return None;
+    }
+    let id = frame[11] as u16;
+    if id == 0 {
+        return None;
+    }
+    // Validate the encoded ID exists in the configured zone map; an
+    // unknown id is a stale or hostile frame. Single hash lookup —
+    // the value isn't needed, just presence.
+    forwarding.zone_id_to_name.get(&id).map(|_| id)
+}
+
+pub(in crate::afxdp) fn parse_packet_destination_from_frame(
+    frame: &[u8],
+    meta: UserspaceDpMeta,
+) -> Option<IpAddr> {
+    let l3 = meta.l3_offset as usize;
+    match meta.addr_family as i32 {
+        libc::AF_INET => {
+            let end = l3.checked_add(20)?;
+            if end > frame.len() {
+                return None;
+            }
+            Some(IpAddr::V4(Ipv4Addr::new(
+                frame[l3 + 16],
+                frame[l3 + 17],
+                frame[l3 + 18],
+                frame[l3 + 19],
+            )))
+        }
+        libc::AF_INET6 => {
+            let end = l3.checked_add(40)?;
+            if end > frame.len() {
+                return None;
+            }
+            Some(IpAddr::V6(Ipv6Addr::from(
+                <[u8; 16]>::try_from(&frame[l3 + 24..l3 + 40]).ok()?,
+            )))
+        }
+        _ => None,
+    }
+}
+
+pub(in crate::afxdp) fn try_parse_metadata(area: &MmapArea, desc: XdpDesc) -> Option<UserspaceDpMeta> {
+    let meta_len = std::mem::size_of::<UserspaceDpMeta>();
+    if (desc.addr as usize) < meta_len {
+        return None;
+    }
+    let meta_offset = (desc.addr as usize).checked_sub(meta_len)?;
+    let bytes = area.slice(meta_offset, meta_len)?;
+    let meta = unsafe { *(bytes.as_ptr() as *const UserspaceDpMeta) };
+    if meta.magic != USERSPACE_META_MAGIC || meta.version != USERSPACE_META_VERSION {
+        return None;
+    }
+    if meta.length as usize != meta_len {
+        return None;
+    }
+    Some(meta)
 }
